+++ usp-core-waap-csrf-protection/intro/foreground.sh
@@ -1,2 +1,7 @@
 #!/bin/bash
+
+# SPDX-FileCopyrightText: 2025 United Security Providers AG, Switzerland
+#
+# SPDX-License-Identifier: GPL-3.0-only
+
 FILE=~/.scenario_staging/intro-foreground.sh; while ! test -f ${FILE}; do clear; sleep 0.1; done; bash ${FILE}