+++ usp-core-waap-csrf-protection/step3/verify.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
 
+# SPDX-FileCopyrightText: 2025 United Security Providers AG, Switzerland
+#
+# SPDX-License-Identifier: GPL-3.0-only
+
 grep '/profile' /var/log/containers/juiceshop-usp-core-waap-*.log