+++ usp-core-waap-intro/step1/verify.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
 
+# SPDX-FileCopyrightText: 2025 United Security Providers AG, Switzerland
+#
+# SPDX-License-Identifier: GPL-3.0-only
+
 kubectl logs juiceshop -n juiceshop |grep loginAdminChallenge