+++ usp-core-waap-openapi-intro/step2/verify.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+# SPDX-FileCopyrightText: 2025 United Security Providers AG, Switzerland
+#
+# SPDX-License-Identifier: GPL-3.0-only
+
 kubectl logs \
   -l app.kubernetes.io/name=usp-core-waap \
   -n petstore \