+++ usp-core-waap-virtual-patch/step2/verify.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 
+# SPDX-FileCopyrightText: 2025 United Security Providers AG, Switzerland
+#
+# SPDX-License-Identifier: GPL-3.0-only
+
 kubectl logs \
   -n prometheus \
   -l app.kubernetes.io/name=usp-core-waap \
   --tail=-1 \
-  |grep "coraza-vm.*/debug/pprof" \
-  | sed -e 's/.* coraza-vm: //' \
+  | grep "\[critical\]\[golang\].*/debug/pprof" \
+  | sed -e 's/\[.*\] {/{/' \
   | jq --exit-status \
   'select(."request.path" == "/debug/pprof" and ."crs.violated_rule".id > 299999 and ."crs.violated_rule".id < 400000)'